Put trackers into new Module
